Add scheduled single-run tasks as a planned feature
This feature is mentioned multiple times in the task manager article
(/midpoint/reference/tasks/task-manager/) but it's not implemented yet.
I added the feature to Planned features so that it's clear it's
something we apparently consider as possible if someone needed it and
"adopted" it. Also, I think it should be here so that it's possible to
"track" the feature, i.e., let readers verify that yes, it's not just an
outdated documentation, it's really not impl'd yet.