delete duplicated chapter "Risk-Based Approach..."
This "sect3" chapter was apparently edited and then 1:1 cloned as
"sect2" chapter in 90feef4b31, just below "IAM and Security"
and before the "Zero-Trust Approach" sub-chapter.

That new location of this chapter seems to make sense,
but the now-confusingly-duplicate "sect3" section needs to be
removed - which is done by this commit.