Use xrefv to avoid 404 on 4.9 as target is 4.10-only
Currently, the default branch is 4.9 and links from the `docs` repo leading to the `midpoint` repo land on the `support-4.9`
branch which does not (and never will) have the target page. Hence the need for `xrefv`. We need to remove the explicit version
in the future (as soon as the go to default branch changes).